tests/requirements.txt
Pillow==10.0.1
pytest-cov==4.1.0
scipy==1.10.1